Add CachedAPI with ResourceStore
The CachedAPI operates like a standard ObjectAPI with the difference
that it adds a caching layer between requests and responses. It achieves
this by watching the set of resources that it is interested in, and
maintaining an internal cache of this set.

Get, List, Watch events are all processed locally against this fully
cached set.

Modifying operations (Update, Patch, Apply, Delete, Create, etc) are
transparently passed through.