Refactor Watcher code
We have now split much of the logic of the original streamWatcher out
into the generic stream.AsyncStream package, which takes care of making
the watch event accessible concurrently. This allows the streamWatcher,
now just called Watcher, to be more clearly focused on Kubernetes
related watch logic.

This separation will also allow us to use the more stream lined
non-concurrent Watcher in the ResourceCache, which may reduce its
reliance on ugly mutex locks.

In addition, the Watcher has been updated such that it now attempts to
gracefully reconnect in the event of a non-erroneous disconnection. This
is normally simply caused by a timeout, which clients are expected to
respond to by reconnecting.