Drop support for name in ResourceCache
It's unusual that one would want to form a cache for one specific
resource (and if you do, it's arguable that the ResourceCache is too
heavy a solution for that).

It is also nicer to do a list as it gives us a more up to date
resourceVersion to watch from, rather than the potentially much older
resourceVersion on an object itself.

If clients really want to use a ResourceCache on a single resource, they
can emulate the functionality by setting the metadata.name
fieldSelector.